+++ test/test_benchmarks.py
@@ -9,6 +9,9 @@ from test.conftest import (
     store_fixture,
     search_records,
     search_terms_fixture,
+    searchable_records_fixture,
+    searchable_keys_fixture,
+    searchable_store_fixture,
 )
 from test.utils import fill_store
 
@@ -19,9 +22,22 @@ def test_benchmark_set(benchmark, store, k, v):
     benchmark(store.set, k=k, v=v)
 
 
+@pytest.mark.parametrize("store, k, v", searchable_records_fixture)
+def test_benchmark_set_with_search(benchmark, store, k, v):
+    """Benchmarks the set operation when search is enabled"""
+    benchmark(store.set, k=k, v=v)
+
+
 @pytest.mark.parametrize("store, k", keys_fixture)
 def test_benchmark_get(benchmark, store, k):
     """Benchmarks the get operation"""
+    fill_store(store=store, data=records)
+    benchmark(store.get, k=k)
+
+
+@pytest.mark.parametrize("store, k", searchable_keys_fixture)
+def test_benchmark_get_with_search(benchmark, store, k):
+    """Benchmarks the get operation when search is enabled"""
     fill_store(store=store, data=records)
     benchmark(store.get, k=k)
 
@@ -47,6 +63,13 @@ def test_benchmark_delete(benchmark, store, k):
     benchmark(store.delete, k=k)
 
 
+@pytest.mark.parametrize("store, k", searchable_keys_fixture)
+def test_benchmark_delete_with_search(benchmark, store, k):
+    """Benchmarks the delete operation when search is enabled"""
+    fill_store(store=store, data=records)
+    benchmark(store.delete, k=k)
+
+
 @pytest.mark.parametrize("store", store_fixture)
 def test_benchmark_clear(benchmark, store):
     """Benchmarks the clear operation"""
@@ -54,8 +77,22 @@ def test_benchmark_clear(benchmark, store):
     benchmark(store.clear)
 
 
+@pytest.mark.parametrize("store", searchable_store_fixture)
+def test_benchmark_clear_with_search(benchmark, store):
+    """Benchmarks the clear operation when search is enabled"""
+    fill_store(store=store, data=records)
+    benchmark(store.clear)
+
+
 @pytest.mark.parametrize("store", store_fixture)
 def test_benchmark_compact(benchmark, store):
     """Benchmarks the compact operation"""
+    fill_store(store=store, data=records)
+    benchmark(store.compact)
+
+
+@pytest.mark.parametrize("store", searchable_store_fixture)
+def test_benchmark_compact_with_search(benchmark, store):
+    """Benchmarks the compact operation when search is enabled"""
     fill_store(store=store, data=records)
     benchmark(store.compact)